requirements.txt
python-telegram-bot>=21.0,<22
aiosqlite>=0.20
aiohttp>=3.9
telegramify-markdown>=0.1.0
